fix: rebalance robots txt generator layout
